refactor: refactored some of the code and edited some styles
Close: #89
