+++ apps/client/src/components/main-page/ChatOverallComponent.tsx
@@ -1,4 +1,4 @@
-import { FC, useState } from 'react';
+import { FC } from 'react';
 import { Outlet } from 'react-router-dom';
 
 import { UserConversationList } from './UserConversationsList';
