+++ apps/client/src/utils/errorSwalMessage.ts
@@ -1,0 +1,9 @@
+import Swal from 'sweetalert2';
+
+export const errorSwalMessage = (error: any) => (
+  Swal.fire({
+    icon: 'error',
+    title: 'Oops...',
+    text: error?.response.data.msg + " " + error?.response.status,
+  })
+);
