+++ apps/client/src/utils/index.ts
@@ -1,2 +1,5 @@
 export * from './formatTime';
 export * from './getPreSignedPutUrl';
+export * from './services/apiClient';
+export * from './errorSwalMessage';
+export * from './successSwalMessage';
