+++ apps/client/src/utils/services/apiClient.ts
@@ -1,0 +1,8 @@
+import axios from 'axios';
+
+import { baseUrl } from '../../config/environment';
+
+export const apiClient = axios.create({
+  baseURL: `${baseUrl}/api/v1`,
+  withCredentials: true,
+});
