+++ apps/client/src/utils/successSwalMessage.ts
@@ -1,0 +1,11 @@
+import Swal from 'sweetalert2';
+
+export const successSwalMessage = (msg: string) => (
+  Swal.fire({
+    position: 'top-end',
+    icon: 'success',
+    title: msg,
+    showConfirmButton: false,
+    timer: 1500
+  })
+);
