+++ apps/server/src/__tests__/getPresignedUrl.test.ts
@@ -15,10 +15,6 @@ describe('Testing the route of creating a presigned url to upload media to space
       .set('Cookie', [`token=${mustafaToken}`])
       .expect(201);
 
-    console.log(typeof res.body.data.url, 'Presigned URL data type!!');
-    console.log(res.body.data.url, 'Presigned URL data data!!');
-    console.log(res.body.data, 'Presigned URL body data!!');
-
     expect(typeof res.body.data.url).toBe('string');
     expect(res.body.data.url.startsWith('https://')).toBe(true);
     expect(res.body.msg).toBeDefined();
